Fix Pull-To-Refresh on iOS11.
Pull-to-refresh triggers reloading immediately even pulling down for a
very short distance then releasing on iOS 11.  This PR fixes it.

The root cause is that safe area insets are introduced in iOS 11.  We
should respect it as the contentInsets doesn't seem to adjust anymore.

If we're going to use the old way to adjust contentInsets,
contentInsetAdjustmentBehavior should set to
UIScrollViewContentInsetAdjustmentNever, then adjust contentInsets in
the old way.  Pull-To-Refresh will just work as expected.